requirements.txt
streamlit==1.15.0
Pillow==8.2.0
tensorflow==2.7.0  
numpy==1.21.5
